backend/requirements.txt
asgiref==3.4.1
astroid==2.5
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
coverage==5.3
coveralls==3.3.0
distlib==0.3.2
Django==3.2.6
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.0.12
idna==3.0
importlib-metadata==4.8.0
isort==5.9.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.4.0
platformdirs==2.3.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.7.2
wrapt==1.12.1
zipp==3.6.0
